+++ src/otx/algorithms/detection/adapters/mmdet/task.py
@@ -178,6 +178,7 @@ class MMDetectionTask(OTXDetectionTask):
             model_classes,
             self._input_size,
             train_dataset=train_dataset,
+            max_num_detections=self.max_num_detections,
         )
         if should_cluster_anchors(self._recipe_cfg):
             if train_dataset is not None:
@@ -486,6 +487,12 @@ class MMDetectionTask(OTXDetectionTask):
         assert len(self._precision) == 1
         export_options["precision"] = str(self._precision[0])
         export_options["type"] = str(export_format)
+        if self.max_num_detections > 0:
+            logger.info(f"Export max_num_detections: {self.max_num_detections}")
+            post_proc_cfg = export_options["deploy_cfg"]["codebase_config"]["post_processing"]
+            post_proc_cfg["max_output_boxes_per_class"] = self.max_num_detections
+            post_proc_cfg["keep_top_k"] = self.max_num_detections
+            post_proc_cfg["pre_top_k"] = self.max_num_detections * 10
 
         export_options["deploy_cfg"]["dump_features"] = dump_features
         if dump_features: