+++ src/otx/algorithms/detection/configs/rotated_detection/efficientnetb2b_maskrcnn/model.py
@@ -7,7 +7,7 @@
 # pylint: disable=invalid-name
 
 _base_ = [
-    "../../../../../recipes/stages/instance-segmentation/incremental.py",
+    "../../../../../recipes/stages/instance_segmentation/incremental.py",
     "../../../../common/adapters/mmcv/configs/backbones/efficientnet_b2b.yaml",
     "../../base/models/detector.py",
 ]