+++ tools/experiment.py
@@ -6,6 +6,7 @@
 import argparse
 import csv
 import dataclasses
+import gc
 import json
 import os
 import re
@@ -31,6 +32,7 @@ def get_parser() -> argparse.ArgumentParser:
     parser = argparse.ArgumentParser()
     parser.add_argument("-f", "--file", type=str, help="Experiment recipe file.")
     parser.add_argument("-p", "--parse", type=str, help="Workspace path to parse.")
+    parser.add_argument("-d", "--dryrun", action="store_true", help="Print experiment commands without execution.")
     return parser
 
 
@@ -566,9 +568,9 @@ class ExpRecipeParser:
     ) -> List[Dict[str, str]]:
         if isinstance(target_str, str):
             target_str = [target_str]
-        found_keys = []
+        found_keys = set()
         for each_str in target_str:
-            found_keys.extend([x for x in set(self._replace_pat.findall(each_str)) if x in variable])
+            found_keys.update([x for x in set(self._replace_pat.findall(each_str)) if x in variable])
         if not found_keys:
             return []
 
@@ -594,9 +596,11 @@ class ExpRecipeParser:
             for key, val in variable.items():
                 target = target.replace(f"${{{key}}}", val)
         elif isinstance(target, list):
+            target = target.copy()
             for i in range(len(target)):
                 target[i] = self._replace_var_in_target(variable, target[i])
         elif isinstance(target, dict):
+            target = target.copy()
             for key in target.keys():
                 target[key] = self._replace_var_in_target(variable, target[key])
         else:
@@ -664,7 +668,10 @@ class OtxCommandRunner:
         repeat_idx (int): repeat index.
     """
 
-    OUTPUT_FILE_NAME = {"export": "openvino.bin", "optimize": "weights.pth"}
+    OUTPUT_FILE_NAME: Dict[str, List[str]] = {
+        "export": ["openvino.bin"],
+        "optimize": ["weights.pth", "openvino.bin"]
+    }
 
     def __init__(self, command_ins: Command, repeat_idx: int):
         self._command_ins = command_ins
@@ -673,35 +680,46 @@ class OtxCommandRunner:
         self._workspace = Path("_".join(self._command_var.values()).replace("/", "_") + f"_repeat_{repeat_idx}")
         self._command_var["repeat"] = str(repeat_idx)
         self._fail_logs: List[CommandFailInfo] = []
-        self._previous_cmd_entry: Optional[str] = None
+        self._previous_cmd_entry: Optional[List[str]] = []
 
     @property
     def fail_logs(self) -> List[CommandFailInfo]:
         """Information of all failed cases."""
         return self._fail_logs
 
-    def run_command_list(self):
+    def run_command_list(self, dryrun: bool = False):
         """Run all commands and organize experiment results."""
         for command in self._command_ins.command:
             command = command.split()
-            if not self._prepare_run_command(command):
+            if not self._prepare_run_command(command) and not dryrun:
                 print(f"otx {command[1]} is skipped.")
                 continue
 
-            self._run_otx_command(command)
+            if not dryrun:
+                self._run_otx_command(command)
+            else:
+                print(" ".join(command))
 
-            self._previous_cmd_entry = command[1]
+            self._previous_cmd_entry.append(command[1])
 
-        organize_exp_result(self._workspace, self._command_var)
+            gc.collect()
+
+        if not dryrun:
+            organize_exp_result(self._workspace, self._command_var)
 
     def _prepare_run_command(self, command: List[str]) -> bool:
         self.set_arguments_to_cmd(command, "--workspace", str(self._workspace))
         cmd_entry = command[1]
+        previous_cmd = None
+        for previous_cmd in reversed(self._previous_cmd_entry):
+            if previous_cmd != "eval":
+                break
+
         if cmd_entry == "train":
             self.set_arguments_to_cmd(command, "--seed", str(self._repeat_idx))
         elif cmd_entry == "eval":
-            if self._previous_cmd_entry in self.OUTPUT_FILE_NAME:
-                file_path = self._find_model_path(self._previous_cmd_entry)
+            if previous_cmd in ["export", "optimize"]:
+                file_path = self._find_model_path(previous_cmd)
                 if file_path is None:
                     return False
                 self.set_arguments_to_cmd(command, "--load-weights", str(file_path))
@@ -709,6 +727,12 @@ class OtxCommandRunner:
             else:
                 output_path = str(self._workspace / "outputs" / "latest_trained_model")
             self.set_arguments_to_cmd(command, "--output", output_path)
+        elif cmd_entry == "optimize":
+            if previous_cmd == "export":  # execute PTQ. If not, execute QAT
+                file_path = self._find_model_path(previous_cmd)
+                if file_path is None:
+                    return False
+                self.set_arguments_to_cmd(command, "--load-weights", str(file_path))
 
         return True
 
@@ -724,11 +748,13 @@ class OtxCommandRunner:
         if output_dir is None:
             print(f"There is no {cmd_entry} output directory.")
             return None
-        file_path = list(output_dir.rglob(self.OUTPUT_FILE_NAME[cmd_entry]))
-        if not file_path:
-            print(f"{self.OUTPUT_FILE_NAME[cmd_entry]} can't be found.")
-            return None
-        return file_path[0]
+        for file_name in self.OUTPUT_FILE_NAME[cmd_entry]:
+            file_path = list(output_dir.rglob(file_name))
+            if file_path:
+                return file_path[0]
+
+        print(f"{', '.join(self.OUTPUT_FILE_NAME[cmd_entry])} can't be found.")
+        return None
 
     @staticmethod
     def set_arguments_to_cmd(command: List[str], key: str, value: Optional[str] = None, before_params: bool = True):
@@ -755,11 +781,12 @@ class OtxCommandRunner:
         command.insert(index, key)
 
 
-def run_experiment_recipe(recipe_file: Union[str, Path]):
+def run_experiment_recipe(recipe_file: Union[str, Path], dryrun: bool = False):
     """Run experiments based on the recipe.
 
     Args:
         recipe_file (Union[str, Path]): Recipe file to run.
+        dryrun (bool, optional): Whether to only print experiment commands. Defaults to False.
     """
     exp_recipe = ExpRecipeParser(recipe_file)
     output_path = exp_recipe.output_path
@@ -771,7 +798,7 @@ def run_experiment_recipe(recipe_file: Union[str, Path]):
     for command_ins in exp_recipe.commands:
         for repeat_idx in range(exp_recipe.repeat):
             otx_cmd_runner = OtxCommandRunner(command_ins, repeat_idx)
-            otx_cmd_runner.run_command_list()
+            otx_cmd_runner.run_command_list(dryrun)
             fail_cases.extend(otx_cmd_runner.fail_logs)
 
     os.chdir(current_dir)
@@ -779,7 +806,8 @@ def run_experiment_recipe(recipe_file: Union[str, Path]):
     if fail_cases:
         log_fail_cases(fail_cases, output_path)
 
-    aggregate_all_exp_result(output_path)
+    if not dryrun:
+        aggregate_all_exp_result(output_path)
 
 
 def main():
@@ -790,7 +818,7 @@ def main():
     if args.file is not None and args.parse is not None:
         print("Please give either --file or --parse argument.")
     elif args.file is not None:
-        run_experiment_recipe(args.file)
+        run_experiment_recipe(args.file, args.dryrun)
     elif args.parse is not None:
         organize_exp_result(args.parse)
     else: